Configuracion de Ecommers pages
