+++ src/Router/AppRoutes.jsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import { Route, Routes } from 'react-router'
-import HomePage from '../pages/HomePage'
-import ShopPage from '../pages/ShopPage'
-import CartPage from '../pages/CartPage'
-import ProductPage from '../pages/ProductPage'
+import HomePage from '../pages/Ecommers-Pages/HomePage'
+import ShopPage from '../pages/Ecommers-Pages/ShopPage'
+import CartPage from '../pages/Ecommers-Pages/CartPage'
+import ProductPage from '../pages/Ecommers-Pages/ProductPage'
 
 export default function AppRoutes() {
   return (
