--- src/components/HomePage/Privacy.jsx
@@ -1,10 +1,0 @@
-export const Privacy = () => {
-  return (
-      <div className="overflow-hidden p-4">
-      <div className="flex justify-between py-4 px-10 border-[2px] font-PrincipalFont p-4 ">
-          <h2> © 2024 Acme Inc. All rights reserved.</h2>
-          <h2>Privacy Policy</h2>
-      </div>
-      </div >
-    )
-  }
