+++ src/components/Layout.jsx
@@ -1,24 +1,27 @@
 import React from 'react';
+import { Link } from 'react-router-dom';
 
 export default function Layout() {
   return (
-    <nav className="flex items-center justify-between p-4 bg-white shadow-md font-PrincipalFont">
+    <nav className="flex items-center justify-between p-4 bg-white shadow-md font-PrincipalFont sticky">
       <div className="flex items-center">
         <svg className="w-8 h-8 mr-4" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-          <path d="M3 22L12 2L21 22H3Z" stroke="black" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round"/>
+          <path d="M3 22L12 2L21 22H3Z" stroke="black" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" />
         </svg>
       </div>
       <div className="flex items-center space-x-6">
-        <a href="#" className="text-gray-700 hover:text-gray-900">Home</a>
-        <a href="#" className="text-gray-700 hover:text-gray-900">Shop</a>
-        <a href="#" className="text-gray-700 hover:text-gray-900">About</a>
-        <a href="#" className="text-gray-700 hover:text-gray-900">Contact</a>
-        <button className="text-gray-700 hover:text-gray-900">
+        <a href="#" className="text-gray-700 hover:text-gray-900">
+          <Link to="/"> Home</Link></a>
+        <a href="#" className="text-gray-700 hover:text-gray-900">
+        <Link to="/shop">Shop</Link></a>
+        {/* <a href="#" className="text-gray-700 hover:text-gray-900">        <Link to="/">About</Link></a> */}
+        {/* <a href="#" className="text-gray-700 hover:text-gray-900">        <Link to="/">Contact</Link></a> */}
+        <button className="text-gray-700 hover:text-gray-900"><Link to="/cart">
           <svg className="w-6 h-6" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-            <path d="M9 22C9.55228 22 10 21.5523 10 21C10 20.4477 9.55228 20 9 20C8.44772 20 8 20.4477 8 21C8 21.5523 8.44772 22 9 22Z" stroke="black" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round"/>
-            <path d="M20 22C20.5523 22 21 21.5523 21 21C21 20.4477 20.5523 20 20 20C19.4477 20 19 20.4477 19 21C19 21.5523 19.4477 22 20 22Z" stroke="black" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round"/>
-            <path d="M1 1H5L7.68 14.39C7.77144 14.8504 8.02191 15.264 8.38755 15.5583C8.75318 15.8526 9.2107 16.009 9.68 16H19.4C19.8693 16.009 20.3268 15.8526 20.6925 15.5583C21.0581 15.264 21.3086 14.8504 21.4 14.39L23 6H6" stroke="black" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round"/>
-          </svg>
+            <path d="M9 22C9.55228 22 10 21.5523 10 21C10 20.4477 9.55228 20 9 20C8.44772 20 8 20.4477 8 21C8 21.5523 8.44772 22 9 22Z" stroke="black" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" />
+            <path d="M20 22C20.5523 22 21 21.5523 21 21C21 20.4477 20.5523 20 20 20C19.4477 20 19 20.4477 19 21C19 21.5523 19.4477 22 20 22Z" stroke="black" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" />
+            <path d="M1 1H5L7.68 14.39C7.77144 14.8504 8.02191 15.264 8.38755 15.5583C8.75318 15.8526 9.2107 16.009 9.68 16H19.4C19.8693 16.009 20.3268 15.8526 20.6925 15.5583C21.0581 15.264 21.3086 14.8504 21.4 14.39L23 6H6" stroke="black" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" />
+          </svg></Link>
         </button>
       </div>
     </nav>
