+++ src/pages/Ecommers-Pages/ProductPage.jsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import { Product } from '../../components/Ecommers-Components/Product/Product'
+import EcommersLayout from '../../Layouts/EcommersLayout'
+
+export default function Home() {
+  return (
+    <>
+      <EcommersLayout>
+        <Product />
+      </EcommersLayout>
+
+    </>
+  )
+}
