--- src/pages/HomePage.jsx
@@ -1,20 +1,0 @@
-import React from 'react'
-import Header from '../components/HomePage/Header'
-import { Card } from '../components/HomePage/Card'
-import { Page1 } from '../components/HomePage/Page1'
-import { Footer } from '../components/HomePage/Footer'
-import { Privacy } from '../components/HomePage/Privacy'
-import Layout from '../components/HomePage/Layout'
-
-export default function Home() {
-  return (
-    <>
-    <Layout/>
-      <Page1 />
-      <Header />
-      <Card />
-      <Footer/>
-      <Privacy/>
-    </>
-  )
-}
